Refactor Context-related code
