+++ http_prompt/context/__init__.py
@@ -1,0 +1,51 @@
+import six
+
+from http_prompt.utils import smart_quote
+
+
+class Context(object):
+
+    def __init__(self, url=None):
+        self.url = url
+        self.headers = {}
+        self.querystring_params = {}
+        self.body_params = {}
+        self.options = {}
+        self.should_exit = False
+
+    def __eq__(self, other):
+        return (self.url == other.url and
+                self.headers == other.headers and
+                self.options == other.options and
+                self.querystring_params == other.querystring_params and
+                self.body_params == other.body_params and
+                self.should_exit == other.should_exit)
+
+    def copy(self):
+        context = Context(self.url)
+        context.headers = self.headers.copy()
+        context.querystring_params = self.querystring_params.copy()
+        context.body_params = self.body_params.copy()
+        context.options = self.options.copy()
+        context.should_exit = self.should_exit
+        return context
+
+    def update(self, context):
+        if context.url:
+            self.url = context.url
+
+        self.headers.update(context.headers)
+        self.querystring_params.update(context.querystring_params)
+        self.body_params.update(context.body_params)
+        self.options.update(context.options)
+        self.should_exit = self.should_exit
+
+    def httpie_options(context):
+        opts = []
+        for k, v in sorted(six.iteritems(context.options)):
+            opt = k
+            if v is not None:
+                opt += ' ' + smart_quote(v)
+            yield
+            opts.append(opt)
+        return opts
